requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
idna==3.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.27.9
requests==2.27.1
rsa==4.8
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
